+++ internal/pkg/runtime/launcher/native/launcher_linux.go
@@ -977,9 +977,8 @@ func (l *Launcher) setCgroups(instanceName string) error {
 	return nil
 }
 
-// PrepareImage perfoms any image preparation required before execution.
-// This is currently limited to extraction or FUSE mount when using the user namespace,
-// and activating any image driver plugins that might handle the image mount.
+// PrepareImage performs any image preparation required before execution.
+// This is currently limited to extraction or FUSE mount.
 func (l *Launcher) prepareImage(c context.Context, image string) error {
 	if strings.HasPrefix(image, "instance://") {
 		return nil
@@ -987,6 +986,8 @@ func (l *Launcher) prepareImage(c context.Context, image string) error {
 
 	insideUserNs, _ := namespaces.IsInsideUserNamespace(os.Getpid())
 	isUserNs := insideUserNs || l.cfg.Namespaces.User
+	noKernelMount := l.cfg.TmpSandbox || isUserNs || l.cfg.SIFFUSE
+	tryFuse := !l.cfg.TmpSandbox
 
 	img, err := imgutil.Init(image, false)
 	if err != nil {
@@ -1002,20 +1003,20 @@ func (l *Launcher) prepareImage(c context.Context, image string) error {
 	case imgutil.SANDBOX:
 		return nil
 	case imgutil.SQUASHFS:
-		if isUserNs || l.cfg.SIFFUSE {
-			return l.prepareSquashfs(c, img, l.cfg.SIFFUSE)
+		if !l.engineConfig.File.AllowKernelSquashfs || noKernelMount {
+			return l.prepareSquashfs(c, img, tryFuse)
 		}
 		// setuid, kernel squashfs permitted, fuse not requested - no action needed
 		return nil
 	case imgutil.ENCRYPTSQUASHFS:
-		if isUserNs || l.cfg.SIFFUSE {
+		if !l.engineConfig.File.AllowKernelSquashfs || noKernelMount {
 			return fmt.Errorf("encrypted SIF files are only supported in setuid mode, with kernel mounts")
 		}
 		// setuid, kernel squashfs permitted, fuse not requested - no action needed
 		return nil
 	case imgutil.EXT3:
 		if isUserNs || l.cfg.SIFFUSE {
-			return l.prepareExtfs(c, img, l.cfg.SIFFUSE)
+			return l.prepareExtfs(c, img)
 		}
 		// setuid, kernel extfs permitted, fuse not requested - no action needed
 		return nil
@@ -1069,7 +1070,7 @@ func (l *Launcher) prepareSquashfs(ctx context.Context, img *imgutil.Image, tryF
 	return fmt.Errorf("extraction failed: %v", err)
 }
 
-func (l *Launcher) prepareExtfs(_ context.Context, _ *imgutil.Image, _ bool) error {
+func (l *Launcher) prepareExtfs(_ context.Context, _ *imgutil.Image) error {
 	// TODO - Enable fuse2fs handling
 	return fmt.Errorf("extfs images can only be run in setuid mode with kernel extfs mounts enabled")
 }